auto_labeling/Autolabeling/requirements.txt
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.28.5
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.21.5
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==4.0.1
Werkzeug==2.0.2
gunicorn==20.0.4